refactor: Header component decomposition.
