+++ src/pages/favorites-page/favorites-page.tsx
@@ -1,6 +1,6 @@
 import { Helmet } from 'react-helmet-async';
 import { getMockOfferPreviews } from '../../mock/offers-previews-mock';
-import Header from '../../components/header/header';
+import Header from '../../components/header';
 import FavoriteList from '../../components/favorite-list/favorite-list';
 import Footer from '../../components/footer/footer';
