+++ src/pages/main-page/main-page.tsx
@@ -1,7 +1,7 @@
 import { Helmet } from 'react-helmet-async';
 import { getMockOfferPreviews } from '../../mock/offers-previews-mock';
 import { CURRENT_OFFERS_COUNT } from '../../const';
-import Header from '../../components/header/header';
+import Header from '../../components/header';
 import Navigation from '../../components/navigation/navigation';
 import Sort from '../../components/sort/sort';
 import MainMap from '../../components/main-map/main-map';
