+++ src/pages/not-found-page/not-found-page.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
 import { Helmet } from 'react-helmet-async';
-import Header from '../../components/header/header';
+import Header from '../../components/header';
 
 export default function NotFoundPage(): JSX.Element {
   return (
